# ARTIQ Pull Request
## Description of Changes

Speedup computation related to list by manual inlining and removing redundant computation. (The embedding module is crazily slow due to a lot of string operations...)

Tested with two simple programs with large arrays:

Ordinary lists:
```python
from artiq.experiment import *

class Test(EnvExperiment):
    def build(self):
        self.setattr_device("core")
        self.array = list(range(100000))

    @kernel
    def run(self):
        i = 0.0
        for x in self.array:
            i += x
        return i
```

Compile time before and after the patch:
```sh
$ time artiq_compile array_test.py
artiq_compile array_test.py  1.50s user 0.05s system 100% cpu 1.543 total
$ export PYTHONPATH=$HOME/code/artiq:$PYTHONPATH
$ time artiq_compile array_test.py
artiq_compile array_test.py  0.68s user 0.04s system 100% cpu 0.717 total
```

Numpy arrays (tested with 1D and 2D, similar result)

```python
from artiq.experiment import *
import numpy

class Test(EnvExperiment):
    def build(self):
        self.setattr_device("core")
        self.array = numpy.array(list(range(100000)), numpy.float)

    @kernel
    def run(self):
        i = 0.0
        for x in self.array:
            i += x
        return i
```

Compile time before and after the patch:
```sh
$ time artiq_compile array_test.py
artiq_compile array_test.py  6.17s user 0.71s system 99% cpu 6.876 total
$ export PYTHONPATH=$HOME/code/artiq:$PYTHONPATH
$ time artiq_compile array_test.py
artiq_compile array_test.py  1.45s user 0.08s system 99% cpu 1.526 total
```

### Related Issue
- #1380: This is only barely related, as the speedup is not related to type checking and I have no complicated artiq code for profiling. It would be helpful if someone can give some more complicated example for profiling.
- #804: Previously it took 11.38s. Now it takes 1.21s on my machine.
- #1117: Seems not an issue with the current master. Would close this.
- #709: Still an error, but it took 3.76s instead of 6.15s to give the error :).

## Type of Changes

<!-- Leave ONLY the corresponding lines for the applicable type of change: -->
|   | Type |
| ------------- | ------------- |
| ✓  | :bug: Bug fix  |
| ✓  | :hammer: Refactoring  |


## Steps (Choose relevant, delete irrelevant before submitting)

### All Pull Requests

- [x] Use correct spelling and grammar.
- [ ] Close/update issues.

### Code Changes

- [x] Run `flake8` to check code style (follow PEP-8 style). `flake8` has issues with parsing Migen/gateware code, ignore as necessary.
- [x] Test your changes or have someone test them. Mention what was tested and how.

Tested with artiq-fast test scripts:
```
nix-shell -I artiqSrc=$CODE/artiq artiq-fast/shell-dev.nix
```

And tested with `test_embedding`.

### Git Logistics

- [x] Split your contribution into logically separate changes (`git rebase --interactive`). Merge/squash/fixup commits that just fix or amend previous commits. Remove unintended changes & cleanup. See [tutorial](https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase).
- [x] Write short & meaningful commit messages. Review each commit for messages (`git show`). Format:
  ```
  topic: description. < 50 characters total.
  
  Longer description. < 70 characters per line
  ```

### Licensing

See [copyright & licensing for more info](https://github.com/m-labs/artiq/blob/master/CONTRIBUTING.rst#copyright-and-sign-off).
ARTIQ files that do not contain a license header are copyrighted by M-Labs Limited and are licensed under LGPLv3+.